reunion of overly divided sub components
